DebiasingWordEmbeddings/requirements.txt
pytorch>=1.3.1
scipy>=1.3.2
numpy>=1.17.4
jupyter>=1.0.0
matplotlib>=3.1.1
scikit-learn>=0.22.1
requests>=2.22.0
PrettyTable>=0.7.2
gensim>=3.8.1
PTable>=0.9.2
tqdm>=4.32.2